Rename of all references of 'shell' to 'scheduler'. All current start files still run properly.
